+++ src/main/java/nbc/chillguys/nebulazone/infra/websocket/eventListener/WebSocketEventListener.java
@@ -1,4 +1,4 @@
-package nbc.chillguys.nebulazone.infra.websocket;
+package nbc.chillguys.nebulazone.infra.websocket.eventListener;
 
 import org.springframework.context.event.EventListener;
 import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
@@ -8,6 +8,7 @@ import org.springframework.web.socket.messaging.SessionDisconnectEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import nbc.chillguys.nebulazone.application.chat.service.ChatMessageService;
+import nbc.chillguys.nebulazone.infra.websocket.SessionUtil;
 
 @Slf4j
 @Component
